Fix floating IP management
Adding floating IP to the network interface required determining
the interface name; this is done by a script in deploy.yml, but
this approach doesn't persist the IP assignment on reboot. This
removes the IP assignment from deploy.yml and instead assigns
the IP in configuration.nix to `ens3` under the assumtion that
Hetzner generates VMs deterministically, which - coupled with
predictable network interface naming - should result in a
deterministic network interface name.